fix(mock-test): code refactor
